requirements.txt
intercept.py
anyio
prompt_toolkit
colorama
